fix multiple bugs in the implementation of SharedCell.
SharedCell::new() and SharedCell::replace() were allowing their argument
to be dropped, making its reference count one lower than it should be
and resulting in use-after-frees. This is solved by calling
mem::forget() on the argument.

SharedCell::get() was not incrementing the reference count of the
returned Shared<T>, also resulting in a reference count one lower than
it should be and resulting in use-after-frees. This is solved by cloning
the contained Shared<T> and then calling mem::forget() on the original.

Added a test which detects these three bugs.